backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pytesseract
Pillow
httpx
python-multipart
